[REFACTOR] market app 따로 분리
