+++ shop/models.py
@@ -1,17 +1,10 @@
 from django.db import models
 
 from account.models import User
+from market.models import Market
 
 
 # Create your models here.
-class Market(models.Model):
-    market_id = models.BigAutoField(primary_key=True, null=False, unique=True)
-    market_name = models.CharField(max_length=30)
-    street_address = models.CharField(max_length=50)
-    postal_address = models.CharField(max_length=50)
-    has_toilet = models.BooleanField()
-    has_parking = models.BooleanField()
-
 
 class Shop(models.Model):
     shop_id = models.BigAutoField(primary_key=True, null=False, unique=True)
